Post-restructure utils fixes
